requirements.txt
faiss-cpu
flask
openai
pandas
python-dotenv
selenium
sentence-transformers